Add analytics events for prefersDarkMode
