+++ packages/lesswrong/client/logging.ts
@@ -6,6 +6,7 @@ import { captureEvent, AnalyticsUtil, userIdentifiedCallback } from '../lib/anal
 import { browserProperties } from '../lib/utils/browserProperties';
 import { sentryUrlSetting, sentryReleaseSetting, sentryEnvironmentSetting } from '../lib/instanceSettings';
 import { getUserEmail } from "../lib/collections/users/helpers";
+import { userPrefersDarkMode } from "../components/themes/usePrefersDarkMode";
 
 const sentryUrl = sentryUrlSetting.get()
 const sentryEnvironment = sentryEnvironmentSetting.get()
@@ -60,6 +61,7 @@ window.addEventListener('load', ev => {
     url: document.location?.href,
     referrer: document.referrer,
     browserProps: browserProperties(),
+    prefersDarkMode: userPrefersDarkMode(),
     performance: {
       memory: (window as any).performance?.memory?.usedJSHeapSize,
       timeOrigin: window.performance?.timeOrigin,
